curiobot_router_demo_using_mcp/requirements.txt
fastapi
uvicorn
httpx
gradio==6.0.0
openai
python-dotenv
pydantic
mcp
openai-agents
